Add config.yaml, argparse, reorganise some files
