+++ .gitignore
@@ -1,8 +1,5 @@
 /__pycache__
 *.pyc
 
-/secrets
 /parrotbot-env
-/gdrive
-/log_stdout
-/log_stderr
+/config.yaml
